+++ src/Piral.Blazor.Orchestrator/ServiceCollectionExtensions.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Piral.Blazor.Orchestrator.Connector;
 using Piral.Blazor.Orchestrator.Loader;
 using Piral.Blazor.Shared;
 
@@ -10,6 +12,8 @@ public static class ServiceCollectionExtensions
 	public static IServiceCollection AddMicrofrontends<TLoader>(this IServiceCollection services)
 		where TLoader : class, IMfLoaderService
 	{
+		var isEmulator = Environment.GetEnvironmentVariable("PIRAL_BLAZOR_DEBUG_ASSEMBLY") is not null;
+
 		services.AddSingleton<IModuleContainerService, ModuleContainerService>();
 		services.AddSingleton<IComponentActivator, MfComponentActivator>();
 		services.AddSingleton<IMfRepository, MfRepository>();
@@ -17,17 +21,19 @@ public static class ServiceCollectionExtensions
         services.AddSingleton<IMfPackageService, MfPackageService>();
 		services.AddSingleton<IEvents, GlobalEvents>();
 
-		if (Environment.GetEnvironmentVariable("PIRAL_BLAZOR_DEBUG_ASSEMBLY") is not null)
+		if (isEmulator)
         {
             services.AddSingleton<TLoader>();
+			services.TryAddScoped<IMfDebugConnector, MfEmulatorConnector>();
             services.AddSingleton<IMfLoaderService, MfLocalLoaderService<TLoader>>();
         }
 		else
         {
+			services.TryAddScoped<IMfDebugConnector, MfEmptyConnector>();
             services.AddSingleton<IMfLoaderService, TLoader>();
         }
 
-		services.AddSingleton<ISnapshotService, FsNugetSnapshotService>();
+		services.TryAddSingleton<ISnapshotService, FsNugetSnapshotService>();
         services.AddSingleton<IMfComponentService, MfComponentService>();
         services.AddHostedService<MfOrchestrationService>();
 		return services;