+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Piral.Blazor.Server Changelog
 
+## 0.3.0 (tbd)
+
+- Fixed issue with DI using scoped services during debugging in emulator mode
+- Added `PageStyles` and `PageScripts` component to decorate layout
+- Added integration to Piral Inspector protocol for enhanced development
+
 ## 0.2.1 (October 19, 2023)
 
 - Fixed issue with nullables in SDK